Spot lights classes implemented with gradual fade to spot area

+++ PointLight.cpp
@@ -21,9 +21,9 @@ PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue,
 	exponent = exp;
 }
 
-void PointLight::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation,
-	          GLfloat diffuseIntensityLocation, GLfloat positionLocation,
-	          GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation,
+	                      GLuint diffuseIntensityLocation, GLuint positionLocation,
+	                      GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
 	// Ambient
 	glUniform3f(ambientColorLocation, color.x, color.y, color.z);
